# DOCS/api/openapi.yaml
openapi: 3.1.0
info:
  title: Sentinel CCTV Platform
  description: '**Gujarat CCTV Integration Hackathon 2026 — Category 1**


    Model 1 (Camera Registry & GIS Mapping) + Model 2 (Unified Viewing Platform + ANPR).


    Connects to the Sentinel sandbox, reads cameras from `/api/ingest`, provides live HLS/WebRTC viewing,
    ANPR detection indexing with fuzzy plate search, watchlist matching, real-time alerts, and route reconstruction
    on a GIS map.'
  version: 1.0.0
paths:
...
  /api/v1/detections/route/{plate_text}:
    get:
      tags:
      - Detections / ANPR (Model 2)
      summary: Route reconstruction — all sightings of a plate ordered by time
      description: 'Returns sorted sightings + computed speed between consecutive cameras.

        Flags physically impossible transitions (>150 km/h).'
      operationId: plate_route_api_v1_detections_route__plate_text__get
      security:
      - HTTPBearer: []
      parameters:
      - name: plate_text
        in: path
        required: true
        schema:
          type: string
          title: Plate Text
      - name: from_dt
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: From Dt
      - name: to_dt
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: To Dt
      - name: purpose
        in: query
        required: false
        schema:
          type: string
          description: Why — recorded in the audit log
          default: investigation
          title: Purpose
        description: Why — recorded in the audit log
      - name: case_ref
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Case Ref
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/analytics/summary:
    get:
      tags:
      - Analytics & Reports
      summary: Platform-wide dashboard summary
      operationId: summary_api_v1_analytics_summary_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
  /api/v1/analytics/top-plates:
    get:
      tags:
      - Analytics & Reports
      summary: Most frequently detected plates
      operationId: top_plates_api_v1_analytics_top_plates_get
      security:
      - HTTPBearer: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          default: 20
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/analytics/detections-by-hour:
    get:
      tags:
      - Analytics & Reports
      summary: Detections per camera per hour (last 24h)
      operationId: detections_by_hour_api_v1_analytics_detections_by_hour_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
  /api/v1/analytics/report/xlsx:
    get:
      tags:
      - Analytics & Reports
      summary: Output report as XLSX (required submission artefact)
      operationId: download_xlsx_api_v1_analytics_report_xlsx_get
      security:
      - HTTPBearer: []
      parameters:
      - name: from_dt
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: From Dt
      - name: to_dt
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: To Dt
      - name: plate
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Restrict to one plate
          title: Plate
        description: Restrict to one plate
      - name: camera_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Restrict to one camera (native id)
          title: Camera Id
        description: Restrict to one camera (native id)
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 50000
          default: 10000
          title: Limit
      - name: purpose
        in: query
        required: false
        schema:
          type: string
          description: Why — recorded in the audit log
          default: hackathon-demonstration
          title: Purpose
        description: Why — recorded in the audit log
      - name: case_ref
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Case Ref
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/analytics/report/pdf:
    get:
      tags:
      - Analytics & Reports
      summary: Output report as PDF (required submission artefact)
      operationId: download_pdf_api_v1_analytics_report_pdf_get
      security:
      - HTTPBearer: []
      parameters:
      - name: from_dt
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: From Dt
      - name: to_dt
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: To Dt
      - name: plate
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Plate
      - name: camera_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Camera Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 50000
          default: 10000
          title: Limit
      - name: purpose
        in: query
        required: false
        schema:
          type: string
          default: hackathon-demonstration
          title: Purpose
      - name: case_ref
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Case Ref
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/analytics/audit:
    get:
      tags:
      - Analytics & Reports
      summary: Audit trail — who searched or exported what, and why
      operationId: audit_trail_api_v1_analytics_audit_get
      security:
      - HTTPBearer: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          default: 200
          title: Limit
      - name: action
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Action
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/analytics/vehicle/{plate_text}:
    get:
      tags:
      - Analytics & Reports
      summary: Vehicle particulars for a plate (VAHAN adapter)
      description: 'Enrich a plate with owner and vehicle details.


        VAHAN is a closed system with no access route for this prototype, so the

        adapter is contract-first and backed by a documented mock. Every response

        states its `source`; a `source` of "mock" is synthetic and must not be treated

        as authoritative. See DOCS/HLD.md §10 for what changes on credential grant.'
      operationId: vehicle_lookup_api_v1_analytics_vehicle__plate_text__get
      security:
      - HTTPBearer: []
      parameters:
      - name: plate_text
        in: path
        required: true
        schema:
          type: string
          title: Plate Text
      - name: purpose
        in: query
        required: false
        schema:
          type: string
          description: Why — recorded in the audit log
          default: investigation
          title: Purpose
        description: Why — recorded in the audit log
      - name: case_ref
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Case Ref
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/analytics/gap-report:
    get:
      tags:
      - Analytics & Reports
      summary: Coverage gap and ageing-infrastructure analysis
      description: 'Where the estate is thin, and which cameras need attention.


        The problem statement asks for "gap-analysis reports for uncovered zones and

        ageing infrastructure"; both halves are computed from the registry. See

        app/gap_analysis.py for how coverage and condition are determined, and what

        is deliberately reported as unknown rather than assumed healthy.'
      operationId: gap_report_api_v1_analytics_gap_report_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
  /api/v1/analytics/gap-report/xlsx:
    get:
      tags:
      - Analytics & Reports
      summary: Gap-analysis report as XLSX (deliverable)
      operationId: gap_report_xlsx_api_v1_analytics_gap_report_xlsx_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
  /api/v1/analytics/scene:
    post:
      tags:
      - Analytics & Reports
      summary: Ingest one minute of scene counts
      description: 'Write one aggregated bucket.


        Guarded like detection ingest: this is analytics about a public place and,

        while it identifies nobody, it still describes police camera coverage.


        Idempotent on (camera, bucket_start). A worker that restarts mid-minute

        re-sends the bucket it was building, and re-running an analysis over

        recorded footage should correct the row rather than duplicate it — which is

        also why the table carries a unique constraint rather than relying on this

        check alone.'
      operationId: ingest_scene_observation_api_v1_analytics_scene_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SceneObservationIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    SceneObservationIn:
      properties:
        camera_native_id:
          type: string
          title: Camera Native Id
        bucket_start:
          type: string
          format: date-time
          title: Bucket Start
        bucket_seconds:
          type: integer
          title: Bucket Seconds
          default: 60
        frames_sampled:
          type: integer
          title: Frames Sampled
          default: 0
        counts_by_label:
          additionalProperties:
            type: integer
          type: object
          title: Counts By Label
          default: {}
        counts_by_category:
          additionalProperties:
            type: integer
          type: object
          title: Counts By Category
          default: {}
        max_confidence:
          type: number
          title: Max Confidence
          default: 0.0
      additionalProperties: false
      type: object
      required:
      - camera_native_id
      - bucket_start
      title: SceneObservationIn
      description: One aggregated minute from a camera.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer